fix test

+++ tests/core/modules/validate_on_init.py
@@ -1,5 +1,6 @@
 """Module for unit testing validation on initialization."""
 
+import numpy as np
 import pandas as pd
 import pandera as pa
 from pandera.typing import DataFrame
@@ -9,7 +10,7 @@ class ExampleSchema(pa.DataFrameModel):
     class Config:
         coerce = True
 
-    a: int
+    a: np.int64
 
 
 ExampleDataFrame = DataFrame[ExampleSchema]
